cleanup

+++ Board.cpp
@@ -171,7 +171,7 @@ void Board::RandomizeBoard(float alivepct)
 		if (rp <= alivepct)
 		{
 			SetCell(cell, Cell::State::Live);
-			cell.SetAge(ra);
+			cell.SetAge(static_cast<uint16_t>(ra));
 		}
 	}
 	_dirty = 1; // must be dirty, we just randomized it
